+++ updater/software/Inkscape.cs
@@ -154,18 +154,20 @@ namespace updater.software
                     return null;
                 }
 
-                // Search for URL part like '<a href="/gallery/item/33460/inkscape-1.2_2022-05-15_dc2aedaf03-x86_4JZVqba.msi">'.
+                // Search for URL part like '<a href="/gallery/item/33460/inkscape-1.2_2022-05-15_dc2aedaf03-x86_4JZVqba.msi">'
+                //    or without bogus like '<a href="/gallery/item/37365/inkscape-1.2.2_2022-12-09_732a01da63-x86.msi">'.
                 //     or
-                // Search for URL part like '<a href="/gallery/item/33463/inkscape-1.2_2022-05-15_dc2aedaf03-x64_5iRsplS.msi">'.
+                // Search for URL part like '<a href="/gallery/item/33463/inkscape-1.2_2022-05-15_dc2aedaf03-x64_5iRsplS.msi">'
+                //    or without bogus like '<a href="/gallery/item/37366/inkscape-1.2.2_2022-12-09_732a01da63-x64.msi">'.
                 string arch = (bits == "32") ? "x86" : "x64";
                 string dateAndHash = null;
                 {
-                    var reUrl = new Regex("<a href=\"([a-zA-Z0-9\\/]+)/inkscape\\-" + Regex.Escape(newVersion) + "(_2[0-9]{3}\\-[0-9]{2}\\-[0-9]{2}_[0-9a-f]+)\\-" + arch + "(_[a-zA-Z0-9]+)\\.msi\">");
+                    var reUrl = new Regex("<a href=\"([a-zA-Z0-9\\/]+)/inkscape\\-" + Regex.Escape(newVersion) + "(_2[0-9]{3}\\-[0-9]{2}\\-[0-9]{2}_[0-9a-f]+)\\-" + arch + "(_[a-zA-Z0-9]+)?\\.msi\">");
                     Match matchUrl = reUrl.Match(htmlCode);
                     if (!matchUrl.Success)
                         return null;
                     dateAndHash = matchUrl.Groups[2].Value;
-                    string bogus = matchUrl.Groups[3].Value;
+                    string bogus = matchUrl.Groups[3].Success ? matchUrl.Groups[3].Value : "";
                     if (bits == "32")
                         newInfo.install32Bit.downloadUrl = "https://inkscape.org/" + matchUrl.Groups[1].Value + "inkscape-" + newVersion + dateAndHash + "-x86" + bogus + ".msi";
                     else
@@ -174,14 +176,16 @@ namespace updater.software
 
                 // Signature files are given in HTML elements like
                 // <a href="https://media.inkscape.org/media/resources/sigs/inkscape-1.2_2022-05-15_dc2aedaf03-x86.msi_eACymJl.sha256"> or
-                // <a href="https://media.inkscape.org/media/resources/sigs/inkscape-1.2_2022-05-15_dc2aedaf03-x64.msi_p22hkKC.sha256">.
+                // <a href="https://media.inkscape.org/media/resources/sigs/inkscape-1.2.2_2022-12-09_732a01da63-x86.msi.sha256"> (no bogus), or
+                // <a href="https://media.inkscape.org/media/resources/sigs/inkscape-1.2_2022-05-15_dc2aedaf03-x64.msi_p22hkKC.sha256"> or
+                // <a href="https://media.inkscape.org/media/resources/sigs/inkscape-1.2.2_2022-12-09_732a01da63-x64.msi.sha256"> (no bogus).
                 {
                     var reUrl = new Regex("<a href=\"https://media\\.inkscape\\.org/media/resources/sigs/inkscape\\-"
-                        + Regex.Escape(newVersion + dateAndHash) + "\\-" + arch + "\\.msi(_[a-zA-Z0-9]+)\\.sha256\">");
+                        + Regex.Escape(newVersion + dateAndHash) + "\\-" + arch + "\\.msi(_[a-zA-Z0-9]+)?\\.sha256\">");
                     Match matchUrl = reUrl.Match(htmlCode);
                     if (!matchUrl.Success)
                         return null;
-                    string bogus = matchUrl.Groups[1].Value;
+                    string bogus = matchUrl.Groups[1].Success ? matchUrl.Groups[1].Value : "";
                     string signatureUrl = "https://media.inkscape.org/media/resources/sigs/inkscape-"
                         + newVersion + dateAndHash + "-" + arch + ".msi" + bogus + ".sha256";